+++ designer/deployment-manager-api/src/main/scala/pl/touk/nussknacker/engine/api/deployment/simple/SimpleStateStatus.scala
@@ -2,12 +2,29 @@ package pl.touk.nussknacker.engine.api.deployment.simple
 
 import pl.touk.nussknacker.engine.api.deployment.StateStatus.StatusName
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus.defaultActions
-import pl.touk.nussknacker.engine.api.deployment.{ScenarioActionName, StateDefinitionDetails, StateStatus}
+import pl.touk.nussknacker.engine.api.deployment.{
+  DeploymentStatus,
+  NoAttributesDeploymentStatus,
+  NoAttributesStateStatus,
+  ProblemDeploymentStatus,
+  ScenarioActionName,
+  StateDefinitionDetails,
+  StateStatus
+}
 import pl.touk.nussknacker.engine.api.process.VersionId
 
 import java.net.URI
 
 object SimpleStateStatus {
+
+  def fromDeploymentStatus(deploymentStatus: DeploymentStatus): StateStatus = {
+    deploymentStatus match {
+      case noAttributes: NoAttributesDeploymentStatus => NoAttributesStateStatus(noAttributes.name.value)
+      // We assume that all deployment status have default allowedActions. Non-default allowedActions have only
+      // statuses that are not deployment statuses but scenario statuses.
+      case problem: ProblemDeploymentStatus => ProblemStateStatus(problem.description)
+    }
+  }
 
   // Represents general problem.
   final case class ProblemStateStatus(description: String, allowedActions: List[ScenarioActionName] = defaultActions)