+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/newdeployment/DeploymentRepository.scala
@@ -1,9 +1,12 @@
 package pl.touk.nussknacker.ui.process.newdeployment
 
+import cats.implicits.{toFoldableOps, toTraverseOps}
 import db.util.DBIOActionInstances._
-import org.postgresql.util.{PSQLException, PSQLState}
+import pl.touk.nussknacker.engine.api.deployment.{DeploymentStatus, DeploymentStatusName, ProblemDeploymentStatus}
+import pl.touk.nussknacker.engine.api.process.ProcessId
+import pl.touk.nussknacker.engine.newdeployment.DeploymentId
 import pl.touk.nussknacker.ui.db.entity.ProcessEntityData
-import pl.touk.nussknacker.ui.db.{DbRef, NuTables}
+import pl.touk.nussknacker.ui.db.{DbRef, NuTables, SqlStates}
 import pl.touk.nussknacker.ui.process.newdeployment.DeploymentEntityFactory.DeploymentEntityData
 import pl.touk.nussknacker.ui.process.newdeployment.DeploymentRepository.{
   ConflictingDeploymentIdError,
@@ -11,24 +14,28 @@ import pl.touk.nussknacker.ui.process.newdeployment.DeploymentRepository.{
 }
 import slick.jdbc.JdbcProfile
 
-import java.sql.SQLIntegrityConstraintViolationException
+import java.sql.{SQLException, Timestamp}
+import java.time.Clock
 import scala.concurrent.ExecutionContext
 
-class DeploymentRepository(dbRef: DbRef)(implicit ec: ExecutionContext) extends NuTables {
+class DeploymentRepository(dbRef: DbRef, clock: Clock)(implicit ec: ExecutionContext) extends NuTables {
 
   override protected val profile: JdbcProfile = dbRef.profile
 
   import profile.api._
 
+  def getScenarioDeploymentsInNotMatchingStatus(
+      scenarioId: ProcessId,
+      statusNames: Set[DeploymentStatusName]
+  ): DB[Seq[DeploymentEntityData]] = {
+    toEffectAll(deploymentsTable.filter(d => d.scenarioId === scenarioId && !(d.statusName inSet statusNames)).result)
+  }
+
   def saveDeployment(deployment: DeploymentEntityData): DB[Either[ConflictingDeploymentIdError, Unit]] = {
     toEffectAll(deploymentsTable += deployment).asTry.map(
       _.map(_ => Right(()))
         .recover {
-          // for postgres
-          case e: PSQLException if e.getSQLState == PSQLState.UNIQUE_VIOLATION.getState =>
-            Left(ConflictingDeploymentIdError(deployment.id))
-          // for other dbs, e.g. hsql
-          case _: SQLIntegrityConstraintViolationException =>
+          case e: SQLException if e.getSQLState == SqlStates.UniqueViolation =>
             Left(ConflictingDeploymentIdError(deployment.id))
         }
         .get
@@ -45,6 +52,32 @@ class DeploymentRepository(dbRef: DbRef)(implicit ec: ExecutionContext) extends 
         .result
         .headOption
         .map(_.map(DeploymentWithScenarioMetadata.apply _ tupled))
+    )
+  }
+
+  def updateDeploymentStatuses(statusesToUpdate: Map[DeploymentId, DeploymentStatus]): DB[Set[DeploymentId]] = {
+    statusesToUpdate.toList
+      .map { case (id, status) =>
+        toEffectAll(
+          updateDeploymentStatus(id, status).map(updated => if (updated) Set(id) else Set.empty[DeploymentId])
+        )
+      }
+      .sequence
+      .map(_.combineAll)
+      // For the performance reasons it is better to run all updates in the one session, transactionally should enforce it
+      .transactionally
+  }
+
+  def updateDeploymentStatus(id: DeploymentId, status: DeploymentStatus): DB[Boolean] = {
+    val problemDescription = ProblemDeploymentStatus.extractDescription(status)
+    toEffectAll(
+      deploymentsTable
+        .filter(d => d.id === id && (d.statusName =!= status.name || d.statusProblemDescription =!= problemDescription))
+        .map(d => (d.statusName, d.statusProblemDescription, d.statusModifiedAt))
+        .update((status.name, problemDescription, Timestamp.from(clock.instant())))
+        .map { result =>
+          if (result > 0) true else false
+        }
     )
   }